A beautiful exception, just in case

+++ src/main/java/todo/controller/TodoRestController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import todo.exception.TodoNotFoundException;
 import todo.model.Todo;
 import todo.repository.TodoRepository;
 
@@ -24,7 +25,6 @@ public class TodoRestController {
 	@RequestMapping(method = RequestMethod.POST,produces = "application/json")
 	ResponseEntity<?> add(@RequestBody Todo input) {
 		Todo result = todoTaskRepository.save(new Todo(input.title));
-
 		HttpHeaders httpHeaders = new HttpHeaders();
 		httpHeaders.setLocation(ServletUriComponentsBuilder.fromCurrentRequest().path("/todo").buildAndExpand(result.getId()).toUri());
 		return new ResponseEntity<>(result,httpHeaders,HttpStatus.CREATED);
@@ -32,7 +32,8 @@ public class TodoRestController {
 	
 	@RequestMapping(value ="/{id}", method = RequestMethod.PUT,produces = "application/json")
 	ResponseEntity<?> update(@PathVariable Long id,@RequestBody Todo input) {
-		
+		if(todoTaskRepository.findById(id) == null)
+			throw new TodoNotFoundException(id);
 		Todo existing = todoTaskRepository.findById(input.getId());
 		existing.title = input.title;
 		existing.finished = input.finished;
@@ -45,7 +46,8 @@ public class TodoRestController {
 	
 	@RequestMapping(value ="/{id}", method = RequestMethod.DELETE)
 	ResponseEntity<Boolean> delete(@PathVariable Long id) {
-		
+		if(todoTaskRepository.findById(id) == null)
+			throw new TodoNotFoundException(id);
 		todoTaskRepository.delete(id);
 		return new ResponseEntity<Boolean>(Boolean.TRUE ,HttpStatus.OK);
 	}
